Data for the assignment of part 1 (Week 4)
It contains the Chicago Crime Data. The data is classified on the basis of type of crimes ex. Theft, Burglary, Assault,etc. It contains the data containing the neighborhoods (where the crime happened) along with their coordinates. I am interested to find out whether certain crimes are more prevalent in certain areas of Chicago.
This kind of information will help the Chicago Police in taking certain steps in certain areas of the city. For example in areas where theft is prevalent some steps can be implemented to prevent it whereas in areas where Burglary is prevalent, some different steps can be taken.